refactor: migrate some more actions into ConversationActivityV2.kt and only show message request incoming menu if it's an incoming message request
